Add files via upload
The MIST data set of handwritten digits

is too easy (convolutional nets can achieve 99.7%)
is overused.
can not represent modern Computer Vision tasks
Fashion-MNIST consists of 60,000 training images and 10,000 test images. It is a MNIST-like fashion product database. The developers believe MNIST has been overused so they created this as a direct replacement for that dataset. Each image is in greyscale and associated with a label from 10 classes.

Fashion-MNIST is a dataset of Zalando's article images—consisting of a training set of 60,000 examples and a test set of 10,000 examples. Each example is a 28x28 grayscale image, associated with a label from 10 classes.

Fashion-MNIST serves as a direct drop-in replacement for the original MNIST dataset for benchmarking machine learning algorithms. It shares the same image size and structure of training and testing splits.

Size: 30 MB

Number of Records: 70,000 images in 10 classes

Images are 28x28 NumPy arrays, with pixel values ranging between 0 and 255.

Labels are an array of integers, ranging from 0 to 9.

Label Description

T-shirt/top
Trouser
Pullover
Dress
Coat
Sandal
Shirt
Sneaker
Bag
Ankle boot